Refactor Bubbles to be separate files
