+++ components/BubbleChart/Legend.tsx
@@ -1,10 +1,10 @@
 import { FC, useContext } from 'react'
 import { css } from 'pretty-lights'
 import { pxToRem } from '../shared/tokens/spacing'
-import { MiniBubbleSVG } from './Bubble'
 import { WorkerDataContext } from './data/WorkerDataProvider'
 import { blankValue, ColorMap, Column, ShapeOptions } from './data/types'
 import { shapePaths } from './shapes/Shape'
+import { MiniBubbleSVG } from './Bubble/MiniBubble'
 
 const legendList = css`
   list-style-type: none;
