+++ components/BubbleChart/LegendSVG.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
+import { MiniBubbleG } from './Bubble/MiniBubble'
 import { blankValue, ChartOptions, ColorMap } from './data/types'
-import { MiniBubbleG } from './Bubble'
 import { shapePaths } from './shapes/Shape'
 
 const LegendSVG: FC<{ textSize: number; chartOptions: ChartOptions }> = ({
